Use Akka HTTP by default Change to AkkaHttpServerComponents More doc refactoring Use the Netty 4.0 querystringdecoder Change to AkkaHttpServerComponents
