--- documentation/manual/experimental/index.toc
@@ -1,1 +1,0 @@
-AkkaHttpServer:Akka HTTP server backend
