+++ documentation/manual/working/commonGuide/configuration/Configuration.md
@@ -6,7 +6,8 @@ This section explains how to configure your Play application.
 - [[Configuration file syntax and features|ConfigFile]]
 - [[Configuring the application secret|ApplicationSecret]]
 - [[Configuring the JDBC connection pool|SettingsJDBC]]
-- [[Configuring Netty server|SettingsNetty]]
 - [[Configuring Play's thread pools|ThreadPools]]
+- [[Configuring Akka Http Server Backend|SettingsAkkaHttp]]
+- [[Configuring Netty Server Backend|SettingsNetty]]
 - [[Configuring logging|SettingsLogger]]
 - [[Configuring WS SSL|WsSSL]]
