+++ documentation/manual/working/commonGuide/configuration/SettingsNetty.md
@@ -1,7 +1,9 @@
 <!--- Copyright (C) 2009-2017 Lightbend Inc. <https://www.lightbend.com> -->
-# Configuring netty
+# Configuring Netty Server Backend
 
-Play 2's main server is built on top of [Netty](http://netty.io/).
+The Netty server backend is built on top of [Netty](http://netty.io/).
+
+> **NOTE**: The Netty server backend is not the default in 2.6.x, and so must be specifically enabled.
 
 ## Default configuration
 
@@ -11,7 +13,7 @@ Play uses the following default configuration:
 
 ## Configuring transport socket
 
-Native socket transport has higher performance and produces less garbage but are only available on linux 
+Native socket transport has higher performance and produces less garbage but is only available on Linux.
 You can configure the transport socket type in `application.conf`:
 
 ```properties
@@ -24,5 +26,5 @@ play.server {
 
 ## Configuring channel options
 
-The available options are defined in [Netty channel option documentation](http://netty.io/4.0/api/io/netty/channel/ChannelOption.html).
-If you are using native socket transport you can set [additional options](http://netty.io/4.0/api/io/netty/channel/epoll/EpollChannelOption.html).
+The available options are defined in [Netty channel option documentation](http://netty.io/4.1/api/io/netty/channel/ChannelOption.html).
+If you are using native socket transport you can set [additional options](http://netty.io/4.1/api/io/netty/channel/epoll/EpollChannelOption.html).
