+++ documentation/manual/working/commonGuide/index.toc
@@ -4,4 +4,5 @@
 !filters:Built-in HTTP filters
 !Modules:Extending Play with modules
 !database:Databases
+!server:Server Backends
 !production:Deploying your application
