+++ documentation/manual/working/commonGuide/server/index.toc
@@ -1,0 +1,3 @@
+Server:Section contents
+AkkaHttpServer:Play with Akka HTTP Server
+NettyServer:Play with Netty Server
